// src/app/models/security/module-form.model.ts
export interface ModuleForm {
    id: number;
    moduleId: number;
    formId: number;
    moduleName: string;
    formName: string;  
    status: number;
}
export interface CreateModelModuleForm {
    id: number;
    moduleId: number;
    formId: number;
    moduleName: string;
    formName: string;
    status: number;
}